Tic-Tac-Toe Game made with Android Studios
Class project that was created in Android Studios using Java and XML. The app is a Tic-Tac-Toe game that uses the Minimax algoritum, recursively, in order to predict a player’s move and never lose a game. Worst case scenario, the game will end in a draw, but never allow the player to win the game.